+++ src/pyppms/user.py
@@ -2,6 +2,7 @@
 
 from loguru import logger as log
 
+from .billing import PpmsBillingInformation
 from .common import dict_from_single_response
 
 
@@ -16,9 +17,10 @@ class PpmsUser:
         The user's email address.
     phone :str
         The user's phone number.
-    billing_code : str
-        The user's billing code (`bcode`). Note that billing codes in PPMS exist
-        at three levels: project, user, group (with descending priority).
+    billing_info : list(PpmsBillingInformation)
+        All billing information associated to the user. Note that billing codes
+        in PPMS exist at three levels: project, user, group (in descending
+        priority).
     fullname : str
         The full name ("``<LASTNAME> <GIVENNAME>``") of the user in PPMS, falling back
         to the ``username`` attribute if empty.
@@ -40,22 +42,31 @@ class PpmsUser:
         """
         details = dict_from_single_response(response_text, graceful=True)
 
+        self.billing_info = []
 
         self.ppms_group_name = str(details["unitlogin"])
+
+        if str(details["bcode"]):
+            billing_info = PpmsBillingInformation(str(details["bcode"]), "user")
+            self.billing_info.append(billing_info)
+
         self.username = str(details["login"])
         self.email = str(details["email"])
         self.phone = str(details["phone"])
-        self.billing_code = str(details["bcode"])
         self.affiliation = str(details["affiliation"])
         self.active = details["active"]
         self._fullname = f"{details['lname']} {details['fname']}"
 
         log.trace(
             f"PpmsUser initialized: username=[{self.username}], email=[{self.email}], "
-            f"billing_code=[{self.billing_code}], "
             f"ppms_group_name=[{self.ppms_group_name}], fullname=[{self._fullname}], "
             f"active=[{self.active}]"
         )
+        if self.billing_info:
+            infos = ""
+            for info in self.billing_info:
+                infos += f"\n- {str(info)}"
+            log.trace(f"PpmsUser [{self.username}] billing information:{infos}")
 
     @property
     def fullname(self):